Separates classes into files, adds javaDoc comments
