+++ app/src/main/java/azul/Color.java
@@ -1,0 +1,10 @@
+package azul;
+
+public enum Color {
+    White,
+    Red,
+    Green,
+    Blue,
+    Yellow,
+    FirstPlayer
+}
