+++ app/src/main/java/azul/ManufactureDisk.java
@@ -1,0 +1,5 @@
+package azul;
+
+class ManufactureDisk extends TileHolder {
+
+}
