+++ app/src/main/java/azul/Middle.java
@@ -1,0 +1,5 @@
+package azul;
+
+class Middle extends TileHolder {
+
+}
