+++ app/src/main/java/azul/Row.java
@@ -1,0 +1,14 @@
+package azul;
+
+public class Row {
+    private Color color;
+    private byte number;
+
+    public Color getColor() {
+        return color;
+    }
+
+    public byte getNumber() {
+        return number;
+    }
+}
